# optimize cloud account deletes to reduce memory use and processing time especially for accounts with lots of activity