TLDR: For a *very* active account, the old code took ~135 seconds and peaked around ~18 GiB, and new code took ~10 seconds and  peaked around ~0.05 GiB.

![old-code](https://user-images.githubusercontent.com/1472326/165167211-bb50a1ab-3550-466d-b350-2ccf8dd4e8c8.png)
![new-code](https://user-images.githubusercontent.com/1472326/165167213-b9eb7ad3-97a0-474b-8d2b-32f7b5af6b6a.png)
Blue line is the active worker process. Black line is the parent celery worker manager process.

```
# old code:
Task api.tasks.delete_cloud_account[d7db90d7-6e52-4c94-9af1-c8f7c26504ce] succeeded in 135.08046991700002s
# new code:
Task api.tasks.delete_cloud_account[4a292157-936f-4e63-b27d-7118abeb3298] succeeded in 10.343064177999999s
```

https://github.com/cloudigrade/cloudigrade/issues/1156

---

I tested this locally by dropping and migrating a brand new database, synthesizing a very large set of data for one user, calling the `delete_cloud_account` task for its cloud account, running the celery worker in `mprof` to capture memory usage information. I repeated this process for the old code and the new code to compare execution time and memory usage. I was running cloudigrade, psql, and redis on localhost.

Once before testing:

```sh
# initial setup
psql -h localhost -U postgres -c "drop schema public cascade; create schema public;" && \
./cloudigrade/manage.py migrate
./cloudigrade/manage.py runserver

# in another shell while runserver is running
http "http://localhost:8000/internal/api/cloudigrade/v1/syntheticdatarequests/" \
        cloud_type=aws \
        since_days_ago="30" \
        account_count="1" \
        instance_count="30000" \
        image_count="1000" \
        run_count_per_instance_mean="10" \
        image_rhel_chance="1.0" \
        expires_at="$(gdate --date="+7 days" -u "+%Y-%m-%dT%H:%M:%S")"

# wait until the data is ready, then kill the runserver process and dump for future reuse with each test
pg_dump -U postgres -f ~/Desktop/local-large-account-dump.sql
```

Then for each test:

```sh
# reload the database from scratch for consistency
dropdb -U postgres postgres && \
createdb -U postgres postgres && \
psql -U postgres -f ~/Desktop/local-large-account-dump.sql && \
psql -U postgres -c 'vacuum full analyze'

# create the delete_cloud_account task
cat << EOF | ./cloudigrade/manage.py shell
from api.tasks import delete_cloud_account
delete_cloud_account.delay(1)
EOF

# profile the celery worker for the delete_cloud_account queue
mprof run --multiprocess celery \
    --app config --workdir cloudigrade \
    worker --max-tasks-per-child 1 --concurrency 1 --task-events \
    --queues delete_cloud_account >>/tmp/worker-output.txt 2>&1 & \
    tail -f /tmp/worker-output.txt
```

Then to capture the two memory plots:

```sh
mprof plot -t 'new code' -o new-code.png -2
mprof plot -t 'old code' -o old-code.png -1
```

I also spot-checked the database using `select count(1) from` on various related table names to confirm they had many tuples before running the delete task and zero after.